+++ dragodis/constants.py
@@ -1,0 +1,7 @@
+"""
+Stores global constants used in dragodis
+"""
+
+# backend disassembler names
+BACKEND_GHIDRA = "Ghidra"
+BACKEND_IDA = "IDA"